requirements.txt
numpy==2.4.2
packaging==26.0
pillow==12.1.1
pygame==2.6.1
pygame-ce==2.5.6
pygame_gui==0.6.14
python-i18n==0.3.9
setuptools==82.0.0
wheel==0.46.3
